+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/AppliedDirective.java
@@ -5,6 +5,8 @@ package com.graphql_java_generator.plugin.language;
 
 import java.util.Map;
 
+import graphql.language.Value;
+
 /**
  * Contains getter for the attributes of a GraphQL directive, as it has been applied on an item of the GraphQL schema.
  * This directive should be a standard GraphQL directive, or one defined in the schema.
@@ -23,6 +25,6 @@ public interface AppliedDirective {
 	 * @return The map with all arguments values, where the key is the argument's name, and the value is the argument's
 	 *         value
 	 */
-	public Map<String, Object> getArgumentValues();
+	public Map<String, Value<?>> getArgumentValues();
 
 }