+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/CustomTemplates_Client_SpringConfiguration.java
@@ -17,7 +17,7 @@ import com.graphql_java_generator.plugin.PluginMode;
 public class CustomTemplates_Client_SpringConfiguration extends AbstractCustomTemplatesSpringConfiguration {
 
 	public CustomTemplates_Client_SpringConfiguration() {
-		super("allGraphQLCases.graphqls", PluginMode.server, customScalars);
+		super("allGraphQLCases.graphqls", PluginMode.client, customScalars);
 	}
 
 }