+++ graphql-maven-plugin-logic/src/test/resources/templates_personalization/client_DirectiveRegistryInitializer.vm.custom.java
@@ -6,26 +6,26 @@ import com.graphql_java_generator.client.directive.DirectiveLocation;
 import com.graphql_java_generator.client.directive.DirectiveRegistry;
 import com.graphql_java_generator.client.directive.DirectiveRegistryImpl;
 import com.graphql_java_generator.client.request.InputParameter;
-import com.graphql_java_generator.customscalars.CustomScalarRegistry;
 import com.graphql_java_generator.customscalars.CustomScalarRegistryImpl;
 
 public class DirectiveRegistryInitializer {
-	
-	
-	CustomScalarRegistry customScalarRegistry = new CustomScalarRegistryImpl();
 	
 	/**
 	 * Initialization of the {@link DirectiveRegistry} with all known custom scalars, that is with all custom scalars
 	 * defined in the project pom
 	 */
-	public void initDirectiveRegistry() {
+	public static DirectiveRegistry initDirectiveRegistry() {
 		DirectiveRegistry directiveRegistry = new DirectiveRegistryImpl();
 		Directive directive;
 		InputParameter param;
 
 #foreach ($directive in $directives)
+		/////////////////////////////////////////////////////////////////////////////////////////////////////
+		// Creating Directive ${directive.name}
+		/////////////////////////////////////////////////////////////////////////////////////////////////////
 		directive = new Directive();
 		directive.setName("${directive.name}");
+		directive.setPackageName("${pluginConfiguration.packageName}");
 #foreach ($argument in $directive.arguments)
 		param = InputParameter.newHardCodedParameter("${argument.name}", null, ${argument.mandatory},
 #if ($argument.graphQLTypeName == "BigDecimal" || 
@@ -43,7 +43,7 @@ public class DirectiveRegistryInitializer {
 					graphql.Scalars.GraphQL${argument.graphQLTypeName}
 #else
 ## It must be a custom scalar
-					customScalarRegistry.getGraphQLScalarType(${argument.graphQLTypeName})
+					CustomScalarRegistryImpl.customScalarRegistry.getGraphQLScalarType("${argument.graphQLTypeName}")
 #end
 				);
 		directive.getArguments().add(param);
@@ -56,6 +56,7 @@ public class DirectiveRegistryInitializer {
 #end
 
 		DirectiveRegistryImpl.directiveRegistry = directiveRegistry;
+		return directiveRegistry;
 	}
 
 }