+++ graphql-maven-plugin-logic/src/test/resources/templates_personalization/client_GraphQLRequest.vm.custom.java
@@ -1,3 +1,4 @@
+/** This template is custom **/
 package ${pluginConfiguration.packageName};
 
 import java.util.Map;