+++ graphql-maven-plugin-logic/src/test/resources/templates_personalization/object_type.vm.custom.java
@@ -24,7 +24,9 @@ import com.graphql_java_generator.annotation.GraphQLInputParameters;
 import com.graphql_java_generator.annotation.GraphQLInputType;
 import com.graphql_java_generator.annotation.GraphQLNonScalar;
 import com.graphql_java_generator.annotation.GraphQLObjectType;
+import com.graphql_java_generator.annotation.GraphQLQuery;
 import com.graphql_java_generator.annotation.GraphQLScalar;
+import com.graphql_java_generator.annotation.RequestType;
 
 #foreach($import in $imports)
 import $import;
@@ -123,15 +125,15 @@ public class ${targetFileName} #if($object.implementz.size()>0)implements #forea
 #end
 
 		public ${targetFileName} build() {
-			${targetFileName} object = new ${targetFileName}();
+			${targetFileName} _object = new ${targetFileName}();
 #foreach ($field in $object.fields)
 #if(${field.javaName} == '__typename')
-			object.set__typename("${object.javaName}");
+			_object.set__typename("${object.javaName}");
 #else
-			object.set${field.pascalCaseName}(${field.javaName});
+			_object.set${field.pascalCaseName}(${field.javaName});
 #end
 #end
-			return object;
+			return _object;
 		}
 	}
 }